Place meme struct in extra file to separate data
