+++ MemeMe/Meme.swift
@@ -1,0 +1,17 @@
+//
+//  Meme.swift
+//  MemeMe
+//
+//  Created by Heike Bernhard on 26/09/2016.
+//  Copyright © 2016 stufengrau. All rights reserved.
+//
+
+import UIKit
+
+// Data Model for a Meme
+struct Meme {
+    let topText: String
+    let bottomText: String
+    let photo: UIImage
+    let memedPhoto: UIImage
+}
